% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_out.R
\name{ai_ratio}
\alias{ai_ratio}
\title{calc_out}
\usage{
ai_ratio(d, sr, p)
}
\arguments{
\item{d}{Mean subgroup difference of predictor(s)}

\item{sr}{Selection ratio in the full applicant pool}

\item{p}{Proportion of minority group in the full applicant pool}
}
\description{
Helper function to convert mean subgroup differences to AI ratios (Newman et
al., 2007). Called by calc_out().
}
\keyword{internal}
