% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C.R, R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{assert_col_vec}
\alias{assert_col_vec}
\title{assert_col_vec}
\usage{
assert_col_vec(v)

assert_col_vec(v)

assert_col_vec(v)

assert_col_vec(v)

assert_col_vec(v)
}
\arguments{
\item{v}{Intermediate variable v}
}
\value{
Refined variable v

Refined variable v

Refined variable v

Refined variable v

Refined variable v
}
\description{
Support function, refines intermediate variable for use in NBI()

Support function, refines intermediate variable for use in NBI()

Support function, refines intermediate variable for use in NBI()

Support function, refines intermediate variable for use in NBI()

Support function, refines intermediate variable for use in NBI()
}
\keyword{internal}
