% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C.R
\name{assert_col_vec_2C}
\alias{assert_col_vec_2C}
\title{assert_col_vec_2C}
\usage{
assert_col_vec_2C(v)
}
\arguments{
\item{v}{Intermediate variable v}
}
\value{
Refined variable v
}
\description{
Support function, refines intermediate variable for use in NBI()
}
\keyword{internal}
