% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{assert_col_vec_2C_1AIR}
\alias{assert_col_vec_2C_1AIR}
\title{assert_col_vec_2C_1AIR}
\usage{
assert_col_vec_2C_1AIR(v)
}
\arguments{
\item{v}{Intermediate variable v}
}
\value{
Refined variable v
}
\description{
Support function, refines intermediate variable for use in NBI()
}
\keyword{internal}
