% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{assert_col_vec_3C}
\alias{assert_col_vec_3C}
\title{myCon_ineq_3C}
\usage{
assert_col_vec_3C(v)
}
\arguments{
\item{v}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition
}
\keyword{internal}
