% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_out.R
\name{calc_out}
\alias{calc_out}
\title{calc_out}
\usage{
calc_out(x)
}
\arguments{
\item{x}{Matrix of predictor weights solutions}
}
\value{
Expected objective outcomes
}
\description{
Helper function to calculate the expected objective outcome values based on
predictor weights solutions. Called by MOST().
}
\keyword{internal}
