% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{combR}
\alias{combR}
\title{combR}
\usage{
combR(Rx, Ry)

combR(Rx, Ry)

combR(Rx, Ry)

combR(Rx, Ry)
}
\arguments{
\item{Rx}{Predictor inter-correlation matrix}

\item{Ry}{Predictor-criterion correlation (validity)}
}
\value{
Rxy Predictor-criterion correlation matrix

Rxy Predictor-criterion correlation matrix

Rxy Predictor-criterion correlation matrix

Rxy Predictor-criterion correlation matrix
}
\description{
Support function to create predictor-criterion matrix

Support function to create predictor-criterion matrix

Support function to create predictor-criterion matrix

Support function to create predictor-criterion matrix
}
\keyword{internal}
