% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{combR_1C_2AIR}
\alias{combR_1C_2AIR}
\title{combR_1C_2AIR}
\usage{
combR_1C_2AIR(Rx, Ry)
}
\arguments{
\item{Rx}{Predictor inter-correlation matrix}

\item{Ry}{Predictor-criterion correlation (validity)}
}
\value{
Rxy Predictor-criterion correlation matrix
}
\description{
Support function to create predictor-criterion matrix
}
\keyword{internal}
