% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{combR_3C}
\alias{combR_3C}
\title{combR_3C}
\usage{
combR_3C(Rx, Ry)
}
\arguments{
\item{Rx}{Predictor inter-correlation matrix}

\item{Ry}{Predictor-criterion correlation (validity)}
}
\value{
Rxy Predictor-criterion correlation matrix
}
\description{
Support function to create predictor-criterion matrix
}
\keyword{internal}
