% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C.R, R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{dimFun}
\alias{dimFun}
\title{dimFun}
\usage{
dimFun(x)

dimFun(x)

dimFun(x)

dimFun(x)

dimFun(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
x Checked and refined input predictor weight vector

x Checked and refined input predictor weight vector

x Checked and refined input predictor weight vector

x Checked and refined input predictor weight vector

x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x

Support function, checks input predictor weight vector x

Support function, checks input predictor weight vector x

Support function, checks input predictor weight vector x

Support function, checks input predictor weight vector x
}
\keyword{internal}
