% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{dimFun_1C_2AIR}
\alias{dimFun_1C_2AIR}
\title{dimFun_1C_2AIR}
\usage{
dimFun_1C_2AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x
}
\keyword{internal}
