% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{dimFun_2C_1AIR}
\alias{dimFun_2C_1AIR}
\title{dimFun_2C_1AIR}
\usage{
dimFun_2C_1AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x
}
\keyword{internal}
