% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{myCon_eq_2_1C_2AIR}
\alias{myCon_eq_2_1C_2AIR}
\title{myCon_eq_2}
\usage{
myCon_eq_2_1C_2AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Equal constraint condition for use in NBI()
}
\description{
Support function, defines equal constraint condition
}
\keyword{internal}
