% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myCon_eq_3C}
\alias{myCon_eq_3C}
\title{myCon_eq_3C}
\usage{
myCon_eq_3C(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Equal constraint condition for use in NBI()
}
\description{
Support function, defines equal constraint condition
}
\keyword{internal}
