% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{myCon_ineq_1C_2AIR}
\alias{myCon_ineq_1C_2AIR}
\title{myCon_ineq}
\usage{
myCon_ineq_1C_2AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition
}
\keyword{internal}
