% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{myCon_ineq_2C_1AIR}
\alias{myCon_ineq_2C_1AIR}
\title{myCon_ineq_2C_1AIR}
\usage{
myCon_ineq_2C_1AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition
}
\keyword{internal}
