% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myCon_ineq_3C}
\alias{myCon_ineq_3C}
\title{myCon_ineq_3C}
\usage{
myCon_ineq_3C(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition
}
\keyword{internal}
