% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R
\name{myFM_1C_1AIR}
\alias{myFM_1C_1AIR}
\title{myFM_1C_1AIR}
\usage{
myFM_1C_1AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Supporting function, defines criterion space
}
\keyword{internal}
