% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myFM_3C}
\alias{myFM_3C}
\title{myFM_3C}
\usage{
myFM_3C(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Supporting function, defines criterion space
}
\keyword{internal}
