% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{myLinCom_2C_1AIR}
\alias{myLinCom_2C_1AIR}
\title{myLinCom_2C_1AIR}
\usage{
myLinCom_2C_1AIR(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Support function
}
\keyword{internal}
