% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myLinCom_3C}
\alias{myLinCom_3C}
\title{myLinCom_3C}
\usage{
myLinCom_3C(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Support function
}
\keyword{internal}
