% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C.R, R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{myT}
\alias{myT}
\title{myT}
\usage{
myT(x_t)

myT(x_t)

myT(x_t)

myT(x_t)

myT(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
f Temporary criterion space

f Temporary criterion space

f Temporary criterion space

f Temporary criterion space

f Temporary criterion space
}
\description{
Support function, define criterion space for intermediate step in NBI()

Support function, define criterion space for intermediate step in NBI()

Support function, define criterion space for intermediate step in NBI()

Support function, define criterion space for intermediate step in NBI()

Support function, define criterion space for intermediate step in NBI()
}
\keyword{internal}
