% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{myT_1C_2AIR}
\alias{myT_1C_2AIR}
\title{myT_1C_2AIR}
\usage{
myT_1C_2AIR(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
f Temporary criterion space
}
\description{
Support function, define criterion space for intermediate step in NBI()
}
\keyword{internal}
