% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{myT_2C_1AIR}
\alias{myT_2C_1AIR}
\title{myT_2C_1AIR}
\usage{
myT_2C_1AIR(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
f Temporary criterion space
}
\description{
Support function, define criterion space for intermediate step in NBI()
}
\keyword{internal}
