% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{myTCon_eq_2_1C_2AIR}
\alias{myTCon_eq_2_1C_2AIR}
\title{myTCon_eq_2_1C_2AIR}
\usage{
myTCon_eq_2_1C_2AIR(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
ceq Temporary constraint condition
}
\description{
Support function, define constraint condition for intermediate step in NBI()
}
\keyword{internal}
