% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myTCon_eq_3C}
\alias{myTCon_eq_3C}
\title{myTCon_eq_3C}
\usage{
myTCon_eq_3C(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
ceq Temporary constraint condition
}
\description{
Support function, define constraint condition for intermediate step in NBI()
}
\keyword{internal}
