% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R, R/ParetoR_3C.R
\name{myTCon_ineq}
\alias{myTCon_ineq}
\title{myTCon_ineq}
\usage{
myTCon_ineq(x_t)

myTCon_ineq(x_t)
}
\arguments{
\item{x_t}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()

Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition

Support function, defines inequal constraint condition
}
\keyword{internal}
