% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{myTCon_ineq_3C}
\alias{myTCon_ineq_3C}
\title{myTCon_ineq_3C}
\usage{
myTCon_ineq_3C(x_t)
}
\arguments{
\item{x_t}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition
}
\keyword{internal}
