% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R
\name{ParetoR}
\alias{ParetoR}
\title{ParetoR}
\usage{
ParetoR(Rx, Rxy1, sr, prop1, d1, Spac = 10, graph = FALSE)
}
\arguments{
\item{Rx}{Matrix with intercorrelations among predictors}

\item{Rxy1}{Vector with correlation between each predictor and criterion 1}

\item{sr}{Selection ratio in the full applicant pool}

\item{prop1}{Proportion of minority applicants in full applicant pool}

\item{d1}{Subgroup difference; standardized mean differences between minority
and majority subgroups on each predictor in full applicant pool}

\item{Spac}{Number of solutions}

\item{graph}{If TRUE, plots will be generated for Pareto-optimal curve and
predictor weights}
}
\value{
out Pareto-Optimal solution with objective outcome values (Criterion)
and predictor weights (ParetoWeights)
}
\description{
Command function to optimize 1 non-adverse impact objective and 1 adverse
impact objective via NBI algorithm
}
\keyword{internal}
