% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_2AIR.R
\name{ParetoR_1C_2AIR}
\alias{ParetoR_1C_2AIR}
\title{ParetoR_1C_2AIR}
\usage{
ParetoR_1C_2AIR(sr, prop1, prop2, Rx, Rxy1, d1, d2, Spac = 10)
}
\arguments{
\item{sr}{Selection ratio in the full applicant pool}

\item{prop1}{Proportion of minority1 applicants in the full applicant pool}

\item{prop2}{Proportion of minority2 applicants in the full applicant pool}

\item{Rx}{Matrix with intercorrelations among predictors}

\item{Rxy1}{Vector with correlation between each predictor and the non-adverse
impact objective}

\item{d1}{Subgroup difference 1; standardized mean differences between minority1
and majority subgroups on each predictor in full applicant pool}

\item{d2}{Subgroup difference 2; standardized mean differences between minority2
and majority subgroups on each predictor in full applicant pool}

\item{Spac}{Number of solutions}
}
\value{
out Pareto-Optimal solution with objective outcome values (Criterion) and
  predictor weights (ParetoWeights)
}
\description{
Command function to optimize 1 non-adverse impact objective and 2 adverse
impact objectives via NBI algorithm
}
