% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C.R
\name{ParetoR_2C}
\alias{ParetoR_2C}
\title{ParetoR_2C}
\usage{
ParetoR_2C(Rx, Rxy1, Rxy2, Spac = 10, graph = TRUE)
}
\arguments{
\item{Rx}{Matrix with intercorrelations among predictors}

\item{Rxy1}{Vector with correlation between each predictor and non-adverse impact objective 1}

\item{Rxy2}{Vector with correlation between each predictor and non-adverse impact objective 2}

\item{Spac}{Number of Pareto points}

\item{graph}{If TRUE, plots will be generated for Pareto-optimal curve and
predictor weights}
}
\value{
out Pareto-Optimal solution with objective outcome values (Criterion) and
  predictor weights (ParetoWeights)
}
\description{
Command function to optimize 2 non-adverse impact objectives via NBI
algorithm
}
