% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{ParetoR_2C_1AIR}
\alias{ParetoR_2C_1AIR}
\title{ParetoR_2C_1AIR}
\usage{
ParetoR_2C_1AIR(Rx, Rxy1, Rxy2, sr, prop1, d1, Spac = 10)
}
\arguments{
\item{Rx}{Matrix with intercorrelations among predictors}

\item{Rxy1}{Vector with correlation between each predictor and non-adverse
impact objective 1}

\item{Rxy2}{Vector with correlation between each predictor and non-adverse
impact objective 2}

\item{sr}{Selection ratio in full applicant pool}

\item{prop1}{Proportion of minority applicants in full applicant pool}

\item{d1}{Subgroup difference; standardized mean differences between minority
and majority subgroups on each predictor in full applicant pool}

\item{Spac}{Number of Pareto points}
}
\value{
out Pareto-Optimal solution with objective outcome values (Criterion)
  and predictor weights (ParetoWeights)
}
\description{
Command function to optimize 2 non-adverse impact objectives and 1 adverse
impact objective via NBI algorithm
}
