% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_3C.R
\name{ParetoR_3C}
\alias{ParetoR_3C}
\title{ParetoR_3C}
\usage{
ParetoR_3C(Rx, Rxy1, Rxy2, Rxy3, Spac = 10)
}
\arguments{
\item{Rx}{Matrix with intercorrelations among predictors}

\item{Rxy1}{Vector with correlation between each predictor and non-adverse impact objective 1}

\item{Rxy2}{Vector with correlation between each predictor and non-adverse impact objective 2}

\item{Rxy3}{Vector with correlation between each predictor and non-adverse impact objective 3}

\item{Spac}{Number of solutions}
}
\value{
out Pareto-Optimal solution with objective outcome values (Criterion) and
predictor weights (ParetoWeights)
}
\description{
Command function to optimize 3 non-adverse impact objectives via NBI
algorithm
}
\keyword{internal}
