% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R
\name{plotPareto_1C_1AIR}
\alias{plotPareto_1C_1AIR}
\title{plotPareto_1C_1AIR}
\usage{
plotPareto_1C_1AIR(CriterionOutput, ParetoWeights)
}
\arguments{
\item{CriterionOutput}{Pareto-Optimal criterion solution}

\item{ParetoWeights}{Pareto-Optimal predictor weight solution}
}
\value{
Plot of Pareto-optimal curve and plot of predictor weights
}
\description{
Function for plotting Pareto-optimal curve and predictor weights
}
\keyword{internal}
