% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C.R
\name{plotPareto_2C}
\alias{plotPareto_2C}
\title{plotPareto_2C}
\usage{
plotPareto_2C(CriterionOutput, ParetoWeights)
}
\arguments{
\item{CriterionOutput}{Pareto-Optimal criterion solution}

\item{ParetoWeights}{Pareto-Optimal predictor weight solution}
}
\value{
Plot of Pareto-optimal curve and plot of predictor weights
}
\description{
Function for plotting Pareto-optimal curve and predictor weights
}
\keyword{internal}
