% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C.R, R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{Weight_Generate}
\alias{Weight_Generate}
\title{Weight_Generate}
\usage{
Weight_Generate(n, k)

Weight_Generate(n, k)

Weight_Generate(n, k)

Weight_Generate(n, k)

Weight_Generate(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weight_Generate

Weight_Generate

Weight_Generate

Weight_Generate

Weight_Generate
}
\description{
Function intended to test the weight generation scheme for NBI for > 2 objectives

Function intended to test the weight generation scheme for NBI for > 2 objectives

Function intended to test the weight generation scheme for NBI for > 2 objectives

Function intended to test the weight generation scheme for NBI for > 2 objectives

Function intended to test the weight generation scheme for NBI for > 2 objectives
}
\keyword{internal}
