% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C_1AIR.R
\name{Weight_Generate_2C_1AIR}
\alias{Weight_Generate_2C_1AIR}
\title{Weight_Generate_2C_1AIR}
\usage{
Weight_Generate_2C_1AIR(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weight_Generate_2C_1AIR
}
\description{
Function intended to test the weight generation scheme for NBI for > 2 objectives
}
\keyword{internal}
