% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R, R/ParetoR_1C_2AIR.R,
%   R/ParetoR_2C.R, R/ParetoR_2C_1AIR.R, R/ParetoR_3C.R
\name{WeightsFun}
\alias{WeightsFun}
\title{WeightsFun}
\usage{
WeightsFun(n, k)

WeightsFun(n, k)

WeightsFun(n, k)

WeightsFun(n, k)

WeightsFun(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weights All possible weights for NBI subproblem

Weights All possible weights for NBI subproblem

Weights All possible weights for NBI subproblem

Weights All possible weights for NBI subproblem

Weights All possible weights for NBI subproblem
}
\description{
Support function, generates all possible weights for NBI subproblems

Support function, generates all possible weights for NBI subproblems

Support function, generates all possible weights for NBI subproblems

Support function, generates all possible weights for NBI subproblems

Support function, generates all possible weights for NBI subproblems
}
\keyword{internal}
