% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R
\name{WeightsFun_1C_1AIR}
\alias{WeightsFun_1C_1AIR}
\title{WeightsFun_1C_1AIR}
\usage{
WeightsFun_1C_1AIR(n, k)
}
\arguments{
\item{n}{the number of objectives}

\item{k}{the inverse of the 1/k, which is the unform spacing between two w_i (k integral)}
}
\value{
x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x
}
\keyword{internal}
