% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_2C.R
\name{WeightsFun_2C}
\alias{WeightsFun_2C}
\title{WeightsFun_2C}
\usage{
WeightsFun_2C(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weights All possible weights for NBI subproblem
}
\description{
Support function, generates all possible weights for NBI subproblems
}
\keyword{internal}
