% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR_1C_1AIR.R
\name{WeightsFun_1C_1AIR}
\alias{WeightsFun_1C_1AIR}
\title{dimFun_1C_1AIR}
\usage{
WeightsFun_1C_1AIR(n, k)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x
}
\keyword{internal}
